rename local-draft.ts to draft-persistence.ts
